Be resilent when sending Radarr/Sonarr requests
Sending requests to Radarr and Sonarr currently causes this application
to fail by not handling the exception raised by Faraday when an
unsuccessful status code is received in responses.

This checks the first call (for free disk space) for errors, such that
the API being down does not crash this application.